double

+++ Hometask43/Program.cs
@@ -1,14 +1,14 @@
 ﻿Console.Write("Введите число b1: ");
-int b1 = int.Parse(Console.ReadLine());
+double b1 = double.Parse(Console.ReadLine());
 
 Console.Write("Введите число k1: ");
-int k1 = int.Parse(Console.ReadLine());
+double k1 = double.Parse(Console.ReadLine());
 
 Console.Write("Введите число b2: ");
-int b2 = int.Parse(Console.ReadLine());
+double b2 = double.Parse(Console.ReadLine());
 
 Console.Write("Введите число k2: ");
-int k2 = int.Parse(Console.ReadLine());
+double k2 = double.Parse(Console.ReadLine());
 
 double x = (b2 - b1) / (k1 - k2);
 double y = k1 * x + b1;
